requirements.txt
anthropic
python-dotenv
pandas
requests
beautifulsoup4
feedparser
lxml
matplotlib
seaborn
reportlab
